fix(triggers): address review on webhook trigger abstraction
- github: route an issue_comment on a pull request onto the PR's
  conversation (issue.pull_request marker) instead of splitting it into a
  separate /issue: thread for the same number
- triggers: scope the dispatch dedup workflow id by trigger instance so the
  body-hash fallback event id (used when a vendor sends no delivery id)
  can't collide across instances and have REJECT_DUPLICATE drop a delivery
- assistants: inline the Linear entity data / GitHub raw webhook payload
  into the turn text so the assistant can actually act on the event it is
  instructed to inspect
- slack: anchor a channel-less block_actions interaction (App Home / modal)
  to the acting user instead of a shared team-level conversation
- assistants: split source_adapter.go into per-source files
