+++ server/internal/assistants/source_adapter.go
@@ -1,14 +1,16 @@
 package assistants
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"time"
 
 	"github.com/google/uuid"
 )
 
+// sourceAdapter renders the per-source pieces of an assistant turn: the
+// conversation-context header, the output-channel guidance, the decoded turn
+// text, and the chat-id derivation. Each external source (Slack, Linear,
+// GitHub, cron, wake) plus the dashboard implements one; the concrete adapters
+// live in the source_adapter_<source>.go files alongside this shared core.
 type sourceAdapter interface {
 	ThreadContext(sourceRefJSON []byte) (string, error)
 	OutputChannelGuidance() string
@@ -44,460 +46,4 @@ func getSourceAdapter(kind string) (sourceAdapter, error) {
 		return nil, fmt.Errorf("assistant source %q is not supported", kind)
 	}
 	return adapter, nil
-}
-
-type slackSourceRef struct {
-	TeamID    string `json:"team_id"`
-	ChannelID string `json:"channel_id"`
-	ThreadID  string `json:"thread_id"`
-	UserID    string `json:"user_id,omitempty"`
-}
-
-type slackEventPayload struct {
-	EnvelopeType string `json:"envelope_type,omitempty"`
-	EventType    string `json:"event_type,omitempty"`
-	Subtype      string `json:"subtype,omitempty"`
-	TeamID       string `json:"team_id"`
-	ChannelID    string `json:"channel_id"`
-	ThreadID     string `json:"thread_id"`
-	UserID       string `json:"user_id,omitempty"`
-	InviterID    string `json:"inviter_id,omitempty"`
-	BotID        string `json:"bot_id,omitempty"`
-	AppID        string `json:"app_id,omitempty"`
-	Text         string `json:"text"`
-	Timestamp    string `json:"timestamp,omitempty"`
-	Reaction     string `json:"reaction,omitempty"`
-	ItemUserID   string `json:"item_user_id,omitempty"`
-	ItemChannel  string `json:"item_channel,omitempty"`
-	ItemTs       string `json:"item_ts,omitempty"`
-	ItemType     string `json:"item_type,omitempty"`
-	ActionID     string `json:"action_id,omitempty"`
-	ActionValue  string `json:"action_value,omitempty"`
-	BlockID      string `json:"block_id,omitempty"`
-}
-
-type slackAdapter struct{ deterministicChatIDAdapter }
-
-func (slackAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref slackSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode slack source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: Slack\n")
-	if ref.TeamID != "" {
-		fmt.Fprintf(&b, "TeamID: %s\n", ref.TeamID)
-	}
-	if ref.ChannelID != "" {
-		fmt.Fprintf(&b, "ChannelID: %s\n", ref.ChannelID)
-	}
-	if ref.ThreadID != "" {
-		fmt.Fprintf(&b, "ThreadID: %s\n", ref.ThreadID)
-	}
-	if ref.UserID != "" {
-		fmt.Fprintf(&b, "UserID: %s\n", ref.UserID)
-	}
-	return b.String(), nil
-}
-
-func (slackAdapter) OutputChannelGuidance() string {
-	return `## Slack output preferences
-
-Text responses are not delivered to the user. To communicate, call a Slack post tool (e.g. platform_slack_post_message, platform_slack_post_ephemeral). If no suitable tool is available, the user will not see your reply.
-
-## Deciding whether to respond
-
-Not every Slack message needs a reply. ALWAYS reply when the turn's EventType is "app_mention" or the message clearly addresses you (a direct question to you, a follow-up to your last reply). For ambient thread messages (EventType "message") and other passive events, first evaluate whether a reply adds value. Stay silent — call no Slack post tool — when the message is clearly directed at another human in the thread, when it asks nothing you can help with, or when you would only restate what has already been said. When staying silent, end the turn without posting anything. Never post a message explaining a tool error or announcing your decision to stay silent.
-
-When relaying an "assistant_mcp_auth_required" AuthURL, deliver it to the OWNER (per the Owner entry in your instructions, when recorded), not whoever triggered this turn. Prefer platform_slack_post_ephemeral targeted at the owner's Slack user ID so only they see it, and render the AuthURL as a Block Kit actions block containing a single primary button rather than as plain text. If ephemeral can't reach the owner in the current channel, DM the owner instead — but AuthURL expires, so first ask if they're ready to authenticate now and only then re-attempt the tool call to mint a fresh URL for the DM. If the requester is someone other than the owner, also tell them (without the URL) that the owner has to complete the authentication. If neither ephemeral nor DM can reach the owner, don't post the URL.`
-}
-
-func (slackAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload slackEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode slack event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	if payload.EventType != "" {
-		fmt.Fprintf(&b, "EventType: %s\n", payload.EventType)
-	}
-	if payload.Subtype != "" {
-		fmt.Fprintf(&b, "Subtype: %s\n", payload.Subtype)
-	}
-	if payload.UserID != "" {
-		fmt.Fprintf(&b, "UserID: %s\n", payload.UserID)
-	}
-	if payload.InviterID != "" {
-		fmt.Fprintf(&b, "InviterID: %s\n", payload.InviterID)
-	}
-	if payload.Timestamp != "" {
-		fmt.Fprintf(&b, "Timestamp: %s\n", payload.Timestamp)
-	}
-	if payload.Reaction != "" {
-		fmt.Fprintf(&b, "Reaction: :%s:\n", payload.Reaction)
-	}
-	if payload.ItemUserID != "" {
-		fmt.Fprintf(&b, "ItemUserID: %s\n", payload.ItemUserID)
-	}
-	if payload.ItemChannel != "" {
-		fmt.Fprintf(&b, "ItemChannel: %s\n", payload.ItemChannel)
-	}
-	if payload.ItemTs != "" {
-		fmt.Fprintf(&b, "ItemTs: %s\n", payload.ItemTs)
-	}
-	if payload.ItemType != "" {
-		fmt.Fprintf(&b, "ItemType: %s\n", payload.ItemType)
-	}
-	if payload.ActionID != "" {
-		fmt.Fprintf(&b, "ActionID: %s\n", payload.ActionID)
-	}
-	if payload.BlockID != "" {
-		fmt.Fprintf(&b, "BlockID: %s\n", payload.BlockID)
-	}
-	if payload.ActionValue != "" {
-		fmt.Fprintf(&b, "ActionValue: %s\n", payload.ActionValue)
-	}
-	b.WriteString("</message-context>\n\n")
-	b.WriteString(payload.Text)
-	return b.String(), nil
-}
-
-type linearSourceRef struct {
-	EventType string `json:"event_type,omitempty"`
-	URL       string `json:"url,omitempty"`
-}
-
-type linearEventPayload struct {
-	EventType  string          `json:"event_type,omitempty"`
-	Type       string          `json:"type,omitempty"`
-	Action     string          `json:"action,omitempty"`
-	URL        string          `json:"url,omitempty"`
-	Data       json.RawMessage `json:"data,omitempty"`
-	ReceivedAt string          `json:"received_at,omitempty"`
-}
-
-type linearAdapter struct{ deterministicChatIDAdapter }
-
-func (linearAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref linearSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode linear source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: Linear\n")
-	if ref.EventType != "" {
-		fmt.Fprintf(&b, "EventType: %s\n", ref.EventType)
-	}
-	if ref.URL != "" {
-		fmt.Fprintf(&b, "URL: %s\n", ref.URL)
-	}
-	return b.String(), nil
-}
-
-func (linearAdapter) OutputChannelGuidance() string {
-	return `## Linear output preferences
-
-Text responses are not delivered to Linear. To act on the event (e.g. update the issue, post a comment, change status), call a Linear tool. If no suitable tool is available, log the situation and stop.`
-}
-
-func (linearAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload linearEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode linear event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	if payload.EventType != "" {
-		fmt.Fprintf(&b, "EventType: %s\n", payload.EventType)
-	}
-	if payload.URL != "" {
-		fmt.Fprintf(&b, "URL: %s\n", payload.URL)
-	}
-	if payload.ReceivedAt != "" {
-		fmt.Fprintf(&b, "ReceivedAt: %s\n", payload.ReceivedAt)
-	}
-	b.WriteString("</message-context>\n\n")
-	fmt.Fprintf(&b, "Linear webhook received: %s.", payload.EventType)
-	if payload.URL != "" {
-		fmt.Fprintf(&b, " See %s.", payload.URL)
-	}
-	b.WriteString(" Inspect the event data and act on it as configured.")
-	return b.String(), nil
-}
-
-type githubSourceRef struct {
-	EventType string `json:"event_type,omitempty"`
-	Action    string `json:"action,omitempty"`
-	Repo      string `json:"repo,omitempty"`
-}
-
-type githubEventPayload struct {
-	EventType  string          `json:"event_type,omitempty"`
-	Action     string          `json:"action,omitempty"`
-	Repo       string          `json:"repo,omitempty"`
-	Ref        string          `json:"ref,omitempty"`
-	Number     int             `json:"number,omitempty"`
-	Payload    json.RawMessage `json:"payload,omitempty"`
-	ReceivedAt string          `json:"received_at,omitempty"`
-}
-
-type githubAdapter struct{ deterministicChatIDAdapter }
-
-func (githubAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref githubSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode github source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: GitHub\n")
-	if ref.EventType != "" {
-		fmt.Fprintf(&b, "EventType: %s\n", ref.EventType)
-	}
-	if ref.Action != "" {
-		fmt.Fprintf(&b, "Action: %s\n", ref.Action)
-	}
-	if ref.Repo != "" {
-		fmt.Fprintf(&b, "Repo: %s\n", ref.Repo)
-	}
-	return b.String(), nil
-}
-
-func (githubAdapter) OutputChannelGuidance() string {
-	return `## GitHub output preferences
-
-Text responses are not delivered to GitHub. To act on the event (e.g. comment on an issue/PR, merge a PR, create a review), call a GitHub tool. If no suitable tool is available, log the situation and stop.`
-}
-
-func (githubAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload githubEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode github event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	if payload.EventType != "" {
-		fmt.Fprintf(&b, "EventType: %s\n", payload.EventType)
-	}
-	if payload.Action != "" {
-		fmt.Fprintf(&b, "Action: %s\n", payload.Action)
-	}
-	if payload.Repo != "" {
-		fmt.Fprintf(&b, "Repo: %s\n", payload.Repo)
-	}
-	if payload.Ref != "" {
-		fmt.Fprintf(&b, "Ref: %s\n", payload.Ref)
-	}
-	if payload.Number != 0 {
-		fmt.Fprintf(&b, "Number: %d\n", payload.Number)
-	}
-	if payload.ReceivedAt != "" {
-		fmt.Fprintf(&b, "ReceivedAt: %s\n", payload.ReceivedAt)
-	}
-	b.WriteString("</message-context>\n\n")
-	fmt.Fprintf(&b, "GitHub webhook received: %s", payload.EventType)
-	if payload.Action != "" {
-		fmt.Fprintf(&b, " (%s)", payload.Action)
-	}
-	if payload.Repo != "" {
-		fmt.Fprintf(&b, " on %s", payload.Repo)
-	}
-	b.WriteString(". Inspect the event payload and act on it as configured.")
-	return b.String(), nil
-}
-
-type cronSourceRef struct {
-	TriggerInstanceID string `json:"trigger_instance_id"`
-	Schedule          string `json:"schedule"`
-}
-
-type cronEventPayload struct {
-	Schedule          string `json:"schedule"`
-	FiredAt           string `json:"fired_at"`
-	TriggerInstanceID string `json:"trigger_instance_id"`
-	Note              string `json:"note,omitempty"`
-}
-
-type cronAdapter struct{ deterministicChatIDAdapter }
-
-func (cronAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref cronSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode cron source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: Cron schedule\n")
-	if ref.Schedule != "" {
-		fmt.Fprintf(&b, "Schedule: %s\n", ref.Schedule)
-	}
-	if ref.TriggerInstanceID != "" {
-		fmt.Fprintf(&b, "TriggerInstanceID: %s\n", ref.TriggerInstanceID)
-	}
-	return b.String(), nil
-}
-
-func (cronAdapter) OutputChannelGuidance() string {
-	return `## Cron output preferences
-
-Text responses are not delivered to anyone — no human is watching the cron tick. To produce a side effect or notify someone, call a tool (e.g. post a Slack message, send an email, write to an external system). If no suitable tool is available for the work this tick requires, log the situation and stop.`
-}
-
-func (cronAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload cronEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode cron event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	if payload.FiredAt != "" {
-		fmt.Fprintf(&b, "FiredAt: %s\n", payload.FiredAt)
-	}
-	if payload.Schedule != "" {
-		fmt.Fprintf(&b, "Schedule: %s\n", payload.Schedule)
-	}
-	b.WriteString("</message-context>\n\n")
-	if payload.Schedule != "" {
-		fmt.Fprintf(&b, "Scheduled run fired for schedule %q at %s. Execute the assistant's configured task for this tick.", payload.Schedule, payload.FiredAt)
-	} else {
-		fmt.Fprintf(&b, "Scheduled run fired at %s. Execute the assistant's configured task for this tick.", payload.FiredAt)
-	}
-	if payload.Note != "" {
-		fmt.Fprintf(&b, " Note: %s", payload.Note)
-	}
-	return b.String(), nil
-}
-
-type wakeSourceRef struct {
-	TriggerInstanceID string `json:"trigger_instance_id"`
-	ScheduledAt       string `json:"scheduled_at,omitempty"`
-}
-
-type wakeEventPayload struct {
-	FiredAt           string `json:"fired_at"`
-	ScheduledAt       string `json:"scheduled_at,omitempty"`
-	TriggerInstanceID string `json:"trigger_instance_id"`
-	Note              string `json:"note,omitempty"`
-}
-
-type wakeAdapter struct{ deterministicChatIDAdapter }
-
-func (wakeAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref wakeSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode wake source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: Wake trigger\n")
-	if ref.TriggerInstanceID != "" {
-		fmt.Fprintf(&b, "TriggerInstanceID: %s\n", ref.TriggerInstanceID)
-	}
-	if ref.ScheduledAt != "" {
-		fmt.Fprintf(&b, "ScheduledAt: %s\n", ref.ScheduledAt)
-	}
-	return b.String(), nil
-}
-
-func (wakeAdapter) OutputChannelGuidance() string {
-	return `## Wake output preferences
-
-Text responses are not delivered to anyone — the wake trigger fired against a thread you scheduled earlier. To make progress visible, call a tool (e.g. post to the original channel, send a follow-up message). If no suitable tool is available, log the situation and stop.`
-}
-
-func (wakeAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload wakeEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode wake event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	if payload.FiredAt != "" {
-		fmt.Fprintf(&b, "FiredAt: %s\n", payload.FiredAt)
-	}
-	if payload.ScheduledAt != "" {
-		fmt.Fprintf(&b, "ScheduledAt: %s\n", payload.ScheduledAt)
-	}
-	b.WriteString("</message-context>\n\n")
-	b.WriteString("Wake trigger fired. You scheduled this earlier to resume work in this thread.")
-	if payload.Note != "" {
-		fmt.Fprintf(&b, " Self-note: %s", payload.Note)
-	}
-	return b.String(), nil
-}
-
-type dashboardSourceRef struct {
-	// UserID is the Gram dashboard user driving the conversation (attribution).
-	// The conversation thread is keyed by the caller-supplied correlation id,
-	// not the user id, so a user can start a fresh thread at will.
-	UserID string `json:"user_id"`
-}
-
-type dashboardEventPayload struct {
-	Text   string `json:"text"`
-	UserID string `json:"user_id,omitempty"`
-}
-
-type dashboardAdapter struct{}
-
-func (dashboardAdapter) ThreadContext(sourceRefJSON []byte) (string, error) {
-	var ref dashboardSourceRef
-	if err := json.Unmarshal(sourceRefJSON, &ref); err != nil {
-		return "", fmt.Errorf("decode dashboard source ref: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("## Conversation context\n\n")
-	b.WriteString("Conversation originated on: Gram dashboard\n")
-	if ref.UserID != "" {
-		fmt.Fprintf(&b, "UserID: %s\n", ref.UserID)
-	}
-	return b.String(), nil
-}
-
-func (dashboardAdapter) OutputChannelGuidance() string {
-	return `## Dashboard output preferences
-
-You are answering a Gram user in the web dashboard's side panel. Your reply text is shown to the user directly — just answer in Markdown, conversationally and concisely; prefer compact tables and short summaries over long prose. This is an analyst's side panel, not a chat app.
-
-When relaying an "assistant_mcp_auth_required" AuthURL, render it as a clickable Markdown link in your reply (e.g. ` + "`[Authorize](<AuthURL>)`" + `) — the dashboard reader IS the owner, no tool call is needed.`
-}
-
-// ChatID: the dashboard's correlation key already IS the server-minted chat id
-// (round-tripped by the client). Use it directly; fall back to a deterministic
-// hash if a non-UUID correlation key ever slips through.
-func (dashboardAdapter) ChatID(assistantID uuid.UUID, correlationID string) uuid.UUID {
-	if parsed, err := uuid.Parse(correlationID); err == nil {
-		return parsed
-	}
-	return deterministicChatID(assistantID, correlationID)
-}
-
-func (dashboardAdapter) DecodeTurn(event assistantThreadEventRecord) (string, error) {
-	var payload dashboardEventPayload
-	if err := json.Unmarshal(event.NormalizedPayloadJSON, &payload); err != nil {
-		return "", fmt.Errorf("decode dashboard event payload: %w", err)
-	}
-	var b bytes.Buffer
-	b.WriteString("<message-context>\n")
-	fmt.Fprintf(&b, "EventID: %s\n", event.EventID)
-	// Sourced from the event's immutable created_at, not time.Now(): DecodeTurn
-	// must be byte-stable across retry/replay, or the capture matcher treats the
-	// re-decoded turn as divergent and opens a spurious generation.
-	if !event.CreatedAt.IsZero() {
-		fmt.Fprintf(&b, "Timestamp: %s\n", event.CreatedAt.UTC().Format(time.RFC3339))
-	}
-	if payload.UserID != "" {
-		fmt.Fprintf(&b, "UserID: %s\n", payload.UserID)
-	}
-	b.WriteString("</message-context>\n\n")
-	b.WriteString(payload.Text)
-	return b.String(), nil
 }
