+++ server/internal/background/triggers/webhook_test.go
@@ -8,6 +8,7 @@ import (
 	"encoding/hex"
 	"hash"
 	"net/http"
+	"reflect"
 	"strconv"
 	"testing"
 	"time"
@@ -124,6 +125,52 @@ func TestHMACSchemeBase64SHA1(t *testing.T) {
 		Encoding: "base64",
 	}
 	require.NoError(t, scheme.Verify(body, headers, "shh"))
+}
+
+func TestWebhookFallbackEventIDHashesBody(t *testing.T) {
+	t.Parallel()
+
+	vendor := WebhookVendor{
+		Slug:      "test-vendor",
+		EventType: reflect.TypeFor[struct{}](),
+		Ingest: func(_ []byte, _ http.Header) (*WebhookIngest, error) {
+			// No vendor delivery id: exercises the content-hash fallback.
+			return &WebhookIngest{Event: map[string]any{"ok": true}}, nil
+		},
+	}
+	def := NewWebhookDefinition(vendor, nil, nil, func(map[string]any) (Config, error) { return nil, nil })
+
+	r1, err := def.HandleWebhook([]byte(`{"a":1}`), http.Header{}, nil)
+	require.NoError(t, err)
+	require.NotNil(t, r1.Event)
+	require.NotEmpty(t, r1.Event.EventID)
+
+	// Identical body -> identical fallback id so a genuine redelivery dedups.
+	r2, err := def.HandleWebhook([]byte(`{"a":1}`), http.Header{}, nil)
+	require.NoError(t, err)
+	require.Equal(t, r1.Event.EventID, r2.Event.EventID)
+
+	// Different body -> different id.
+	r3, err := def.HandleWebhook([]byte(`{"a":2}`), http.Header{}, nil)
+	require.NoError(t, err)
+	require.NotEqual(t, r1.Event.EventID, r3.Event.EventID)
+}
+
+func TestTriggerDispatchWorkflowIDScopesByInstance(t *testing.T) {
+	t.Parallel()
+
+	// The content-hash fallback event id is deterministic from the request body
+	// alone, so two unrelated trigger instances receiving an identical body
+	// produce the same event id. The dispatch workflow id must fold in the
+	// instance id so REJECT_DUPLICATE doesn't collapse the two deliveries and
+	// silently drop one.
+	eventID := "same-content-hash"
+	a := triggerDispatchWorkflowID("instance-a", eventID)
+	b := triggerDispatchWorkflowID("instance-b", eventID)
+	require.NotEqual(t, a, b)
+
+	// Redelivery to the same instance still dedups (same workflow id).
+	require.Equal(t, a, triggerDispatchWorkflowID("instance-a", eventID))
 }
 
 func TestHMACSchemeRejectsMissingSignatureHeader(t *testing.T) {
